Add basic sanity testing for end-to-end functionality
